Refactor Device Orientation API
