+++ KDeviceInfo/src/commonMain/kotlin/com/devx/kdeviceinfo/model/android/AndroidInfo.kt
@@ -2,6 +2,7 @@
 
 package com.devx.kdeviceinfo.model.android
 
+import com.devx.kdeviceinfo.model.common.DeviceOrientation
 import com.devx.kdeviceinfo.model.common.Locale
 
 interface AndroidInfo {
