--- KDeviceInfo/src/commonMain/kotlin/com/devx/kdeviceinfo/model/android/DeviceOrientation.kt
@@ -1,12 +1,0 @@
-@file:Suppress("PropertyName")
-
-package com.devx.kdeviceinfo.model.android
-
-interface DeviceOrientation {
-    val UNKNOWN: String
-    val PORTRAIT: String
-    val LANDSCAPE: String
-    val isPortrait: Boolean
-    val isLandscape: Boolean
-    fun getDeviceOrientation(): String
-}
