+++ KDeviceInfo/src/commonMain/kotlin/com/devx/kdeviceinfo/model/common/DeviceOrientation.kt
@@ -1,16 +1,13 @@
 @file:Suppress("PropertyName")
 
-package com.devx.kdeviceinfo.model.ios
+package com.devx.kdeviceinfo.model.common
 
 interface DeviceOrientation {
     val UNKNOWN: String
     val PORTRAIT: String
     val LANDSCAPE: String
-    val PORTRAIT_UPSIDE_DOWN: String
-    val LANDSCAPE_LEFT: String
-    val LANDSCAPE_RIGHT: String
-    val FACE_UP: String
-    val FACE_DOWN: String
     val isPortrait: Boolean
+    val isLandscape: Boolean
+
     fun getDeviceOrientation(): String
 }
